pyproject.toml
[project]
name = "asml-link"
version = "0.1.0"
description = "Python link shortening service with OAuth2 authentication"
readme = "README.md"
requires-python = ">=3.11"
authors = [{name = "Your Name", email = "email@example.com"}]
license = {text = "MIT"}
dependencies = [
    "quart>=0.18.0",
    "sqlalchemy>=2.0",
    "alembic>=1.12.0",
    "authlib>=1.2.0",
    "pyjwt>=2.8.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "aiosqlite>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.9.0",
    "ruff>=0.10.0",
    "pre-commit>=3.4.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]



[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
